refactor code, moved guard syntax checks into their own file
