+++ dcr-engine/index.ts
@@ -1,139 +1,124 @@
 import type {
-      DCRGraph,
-      DataDCR,
-      Marking,
-      SubProcess,
-      EventMap,
-      Event,
-      Trace,
-      EventLog,
-      DCRGraphS,
-      RoleTrace,
-      Nestings,
-      RelationViolations,
-      VariantLog,
-      VariableStore,
-      Value,
-  } from "./src/types";
-import {isSubProcess} from "./src/types";
+  DCRGraph,
+  DataDCR,
+  Marking,
+  SubProcess,
+  EventMap,
+  Event,
+  Trace,
+  EventLog,
+  DCRGraphS,
+  RoleTrace,
+  Nestings,
+  RelationViolations,
+  VariantLog,
+  VariableStore,
+  Value,
+} from "./src/types";
+import { isSubProcess } from "./src/types";
 import {
-      execute,
-      isEnabled,
-      isAccepting,
-      executeS,
-      isEnabledS,
-      isAcceptingS,
-      evaluateGuard,
-      validateGuardSyntax,
-  } from "./src/executionEngine";
-import {moddleToDCR} from "./src/graphConversion";
+  execute,
+  isEnabled,
+  isAccepting,
+  executeS,
+  isEnabledS,
+  isAcceptingS,
+} from "./src/executionEngine";
+import { evaluateGuard, validateGuardSyntax } from "./src/guardEval";
+import { moddleToDCR } from "./src/graphConversion";
 import {
-    copyMarking,
-    parseDurationMs,
-    filterVariantByTopPercentage,
-    filterVariantByBottomPercentage,
-    getBinaryVariants,
-    getVariants,
+  copyMarking,
+  parseDurationMs,
+  filterVariantByTopPercentage,
+  filterVariantByBottomPercentage,
+  getBinaryVariants,
+  getVariants,
 } from "./src/utility";
-import {
-    parseRoleLog,
-    parseNonRoleLog,
-    writeEventLog,
-    parseBinaryLog,
-} from "./src/eventLogs";
-import {
-    replayTraceS,
-    mergeViolations,
-    quantifyViolations,
-} from "./src/conformance";
+import { parseRoleLog, parseNonRoleLog, writeEventLog, parseBinaryLog } from "./src/eventLogs";
+import { replayTraceS, mergeViolations, quantifyViolations } from "./src/conformance";
 import layoutGraph from "./src/layout";
-import {nestDCR} from "./src/nesting";
-import {generateEventLog} from "./src/generation";
+import { nestDCR } from "./src/nesting";
+import { generateEventLog } from "./src/generation";
 import runTest from "./src/tdm";
-import {alignTrace} from "./src/align";
-
+import { alignTrace } from "./src/align";
 import rejectionMiner from "./src/binary";
-
-import mineFromAbstraction, {abstractLog, filter} from "./src/discovery";
-
-import {StringTraceStreamParser} from "./src/parsers/StringTraceStreamParser";
-import {RegexTraceStreamParser} from "./src/parsers/RegexTraceStreamParser";
-import {DOMTraceStreamParser} from "./src/parsers/DOMTraceStreamParser";
-import {StringEventStreamParser} from "./src/parsers/StringEventStreamParser";
-import {RegexEventStreamParser} from "./src/parsers/RegexEventStreamParser";
-import {DOMEventStreamParser} from "./src/parsers/DOMEventStreamParser";
-import {SAXParser} from "./src/parsers/SAXParser";
-import extractGraph,
-{
+import mineFromAbstraction, { abstractLog, filter } from "./src/discovery";
+import { StringTraceStreamParser } from "./src/parsers/StringTraceStreamParser";
+import { RegexTraceStreamParser } from "./src/parsers/RegexTraceStreamParser";
+import { DOMTraceStreamParser } from "./src/parsers/DOMTraceStreamParser";
+import { StringEventStreamParser } from "./src/parsers/StringEventStreamParser";
+import { RegexEventStreamParser } from "./src/parsers/RegexEventStreamParser";
+import { DOMEventStreamParser } from "./src/parsers/DOMEventStreamParser";
+import { SAXParser } from "./src/parsers/SAXParser";
+import extractGraph, {
   type ExtractionResult,
   type ProcessDescription,
   type Mention,
   type Entity,
   type Relation,
-  type ExtractionConfig
+  type ExtractionConfig,
 } from "./src/extraction";
 
 export {
-      type DCRGraph,
-      type DCRGraphS,
-      type DataDCR,
-      type VariableStore,
-      type Value,
-      type EventLog,
-      type EventMap,
-      type Marking,
-      type SubProcess,
-      type Event,
-      type Trace,
-      type RoleTrace,
-      type Nestings,
-      type RelationViolations,
-      type VariantLog,
-      type ExtractionResult,
-      type ProcessDescription,
-      type Mention,
-      type Entity,
-      type Relation,
-      type ExtractionConfig,
-      isSubProcess,
-      execute,
-      isAccepting,
-      isEnabled,
-      moddleToDCR,
-      copyMarking,
-      parseDurationMs,
-      parseRoleLog,
-      parseNonRoleLog,
-      parseBinaryLog,
-      isAcceptingS,
-      executeS,
-      extractGraph,
-      isEnabledS,
-      replayTraceS,
-      writeEventLog,
-      layoutGraph,
-      mineFromAbstraction,
-      abstractLog,
-      nestDCR,
-      filter,
-      mergeViolations,
-      quantifyViolations,
-      getVariants,
-      generateEventLog,
-      runTest,
-      alignTrace,
-      rejectionMiner,
-      getBinaryVariants,
-      filterVariantByTopPercentage,
-      filterVariantByBottomPercentage,
-      StringTraceStreamParser,
-      RegexTraceStreamParser,
-      DOMTraceStreamParser,
-      StringEventStreamParser,
-      RegexEventStreamParser,
-      DOMEventStreamParser,
-      SAXParser,
-      evaluateGuard,
-      validateGuardSyntax,
-  };
-
+  type DCRGraph,
+  type DCRGraphS,
+  type DataDCR,
+  type VariableStore,
+  type Value,
+  type EventLog,
+  type EventMap,
+  type Marking,
+  type SubProcess,
+  type Event,
+  type Trace,
+  type RoleTrace,
+  type Nestings,
+  type RelationViolations,
+  type VariantLog,
+  type ExtractionResult,
+  type ProcessDescription,
+  type Mention,
+  type Entity,
+  type Relation,
+  type ExtractionConfig,
+  isSubProcess,
+  execute,
+  isAccepting,
+  isEnabled,
+  moddleToDCR,
+  copyMarking,
+  parseDurationMs,
+  parseRoleLog,
+  parseNonRoleLog,
+  parseBinaryLog,
+  isAcceptingS,
+  executeS,
+  isEnabledS,
+  extractGraph,
+  replayTraceS,
+  writeEventLog,
+  layoutGraph,
+  mineFromAbstraction,
+  abstractLog,
+  nestDCR,
+  filter,
+  mergeViolations,
+  quantifyViolations,
+  getVariants,
+  generateEventLog,
+  runTest,
+  alignTrace,
+  rejectionMiner,
+  getBinaryVariants,
+  filterVariantByTopPercentage,
+  filterVariantByBottomPercentage,
+  StringTraceStreamParser,
+  RegexTraceStreamParser,
+  DOMTraceStreamParser,
+  StringEventStreamParser,
+  RegexEventStreamParser,
+  DOMEventStreamParser,
+  SAXParser,
+  evaluateGuard,
+  validateGuardSyntax,
+};
